pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "cse"
version = "0.1.0"
description = "Code Synthesis Evolution - A code performance optimization framework"
readme = "README.md"
requires-python = ">=3.10"
license = {file = "LICENSE"}
keywords = ["code-optimization", "llm", "agents"]

classifiers = [
  "Operating System :: OS Independent",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3 :: Only",
]

dependencies = [
    "datasets",
    "numpy",
    "pandas",
    "rich",
    "ruamel.yaml",
    "tenacity",
    "unidiff",
    "simple-parsing",
    "rich-argparse",
    "flask",
    "flask-cors",
    "flask-socketio",
    "pydantic",
    "python-dotenv",
    "pydantic_settings",
    "litellm",
    "GitPython",
    "ghapi",
    "tabulate",
    "textual>=1.0.0",
    "scipy",
    "sortedcontainers",
    "chromadb",
    "pre-commit",
    "PyYAML",
    "tqdm",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "pre-commit",
    "pytest-xdist",
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
namespaces = false

[tool.pytest.ini_options]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]
testpaths = [
    "tests"
]
xfail_strict = true

[tool.ruff]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
    "tests/test_data",
]

line-length = 120
indent-width = 4
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",
    "PLE",
    "E713",
    "E714",
    "E711",
    "F821",
    "F822",
    "F401",
    "F841",
    "F541",
    "F901",
    "I001",
    "I002",
    "UP",
    "C401",
    "C402",
    "C403",
    "C404",
    "C405",
    "F632",
    "W605",
    "B006",
    "B007",
    "B009",
    "EM",
    "FA102",
    "RET",
    "RUF019",
    "PT",
    "SIM201",
    "PTH100",
    "PTH110",
    "PTH111",
    "PTH112",
    "PTH113",
    "PTH114",
    "PTH117",
    "PTH118",
    "PTH119",
    "PTH120",
    "PTH121",
    "PTH122",
    "PTH202",
    "PTH203",
    "PTH204",
    "PTH205",
]
ignore = [
    "RET505",
    "RET506",
    "RET507",
    "E501",
    "E402",
    "E722",
    "E741",
    "PT011",
    "PT018",
]

fixable = ["ALL"]
unfixable = []
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint.isort]
